API Exposure and Setup
This commit modifies config/routes.rb to expose an API under the api namespace. The code there contains a few thoughts on how to expose that api post-deployment

There's also a new getter method on the User model, #walks, that grabs all walks that the User has requested or guarded.

Lastly, there's some setup for the API Users and Walks Controllers