API Allows for Creation of New User
This commit makes changes to the API to allow a new user to be created through HTTP requests.

It creates a namespaced ActionController for the API with the protect_from_forgery setting recommended by the Rails Guide. It sets successful create requests to respond with 201 codes instead of 200 or 203 codes. It also singularizes the private strong param methods in the users and walks controllers.