Implement Materialize Sliding Nav
This commit replaces the old navbar with an options drawer from the
materialize component list. In the layout, it replaces the old nav bar
with this one, slide_bar.

It creates a new js file, main.js which initializes and configures the
sideNav.

Tangentially, it also contains a minor change to the Google Maps script
tag on the walks/new.

Big lesson in this commit is that Rails helpers can take a block which
allows for richer formatting of their content (as seen with the icons in
the menu).